requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
# htmx is a JavaScript library, not a Python package

# Document parsing
pymupdf==1.23.7
python-docx==1.0.1
beautifulsoup4==4.12.2
pdfminer.six==20231228
trafilatura==1.6.2
readability-lxml==0.8.1

# Web crawling
requests==2.31.0

# Vector embeddings and databases
openai==1.3.7
google-cloud-aiplatform==1.36.4
pinecone-client==2.2.4
weaviate-client==3.25.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
tqdm==4.66.1
langchain==0.0.335

# GCP services
google-cloud-storage==2.13.0
google-auth==2.23.4